backend/requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
PyJWT
python-decouple
pydantic[email]